requirements.txt
Django==1.11
djangorestframework==3.7.7
Pygments==2.2.0
django-extensions==2.0.5
ipython==5.5.0
ipython-genutils==0.2.0
httpie==0.9.9
coreapi==2.3.3
coreapi-cli==1.0.9

